Edited the http status being returned

+++ OneOnOne/accounts/views.py
@@ -46,39 +46,6 @@ class LogoutView(generics.DestroyAPIView):
         return self.destroy(request, *args, **kwargs)
 
 
-class EditProfileView(generics.UpdateAPIView):
-    serializer_class = EditProfileSerializer
-    permission_classes = [IsAuthenticated]
-
-    def get_object(self):
-        return self.request.user
-
-    def update(self, request, *args, **kwargs):
-        partial = kwargs.pop('partial', False)
-        instance = self.get_object()
-        original_data = self.get_serializer(instance).data  # Get the original data before update
-
-        serializer = self.get_serializer(instance, data=request.data, partial=partial)
-        serializer.is_valid(raise_exception=True)
-
-        # Handle password update separately
-        if 'password' in serializer.validated_data and serializer.validated_data['password']:
-            instance.set_password(serializer.validated_data['password'])
-            update_session_auth_hash(request, instance)
-
-        self.perform_update(serializer)
-
-        # Return updated profile details along with unchanged fields
-        updated_data = serializer.data
-        response_data = {'detail': 'Profile updated successfully', 'profile': {}}
-
-        # Include unchanged fields from the original profile data
-        for key, value in original_data.items():
-            response_data['profile'][key] = updated_data.get(key, value)
-
-        return Response(response_data, status=status.HTTP_200_OK)
-
-
 class ContactView(APIView):
 
     def get(self, request):
@@ -88,7 +55,6 @@ class ContactView(APIView):
             new = dict(list(x.items())[-2:])
             newContacts.append(new)
 
-        #qs_json = serializers.serialize('json', contacts)
         return Response({"contacts": newContacts})
 
 
@@ -103,10 +69,9 @@ class AddContactView(generics.UpdateAPIView):
         instance = self.get_object()
 
         serializer = self.get_serializer(instance, data=request.data)
-        print(self.request.user)
         if not serializer.is_valid():
             response_data = {'detail': 'Both fields are required'}
-            return Response(response_data)
+            return Response(response_data, status=status.HTTP_400_BAD_REQUEST)
 
         self.perform_update(serializer)
 
@@ -138,23 +103,20 @@ class DeleteContactView(generics.UpdateAPIView):
         serializer = self.get_serializer(instance, data=request.data)
         if not serializer.is_valid():
             response_data = {'detail': 'Email field is required'}
-            return Response(response_data)
+            return Response(response_data, status=status.HTTP_400_BAD_REQUEST)
 
         self.perform_update(serializer)
 
         # Return updated profile details along with unchanged fields
         updated_data = serializer.data
-        response_data = {'detail': 'Contact removed successfully', 'contacts': {}}
+        response_data = {'detail': 'Contact removed successfully'}
 
         email = updated_data['contact_email']
 
-        if not Contact.objects.filter(email=email):
+        if not Contact.objects.filter(owner=instance, email=email):
             response_data = {'detail': 'No contact with that email exists'}
-            return Response(response_data, status=status.HTTP_200_OK)
+            return Response(response_data, status=status.HTTP_400_BAD_REQUEST)
 
-        Contact.objects.filter(email=email).delete()
+        Contact.objects.filter(owner=instance, email=email).delete()
 
         return Response(response_data, status=status.HTTP_200_OK)
-
-# if User.objects.exclude(pk=self.get_serializer(instance).pk).filter(username="nealon").exists():
-#     print("user exists")
